.history/services/inference_service/requirements_20250422144849.txt
fastapi>=0.111
uvicorn[standard]>=0.29
torch==2.2.2
torchvision==0.17.2
pillow>=10.2
pytest
httpx # FastAPI TestClient's async transport
